Think I get anim now. Also have new plan for output streams
The goal is to have seperate files (streams) for primtive/fbx types
and then have a graph-construction stream. This will be a record of
the walk and will basically be a stream of 'opcodes' you use to
construct the graph. So by 'playing back' this record you make the
full graph.

The work we did so far is very important as there is a bunch of stuff
that isnt queryable, however all the properties are, so we can happily
drop those.

The reason for this change was that, when I tried to think of an easy
way around the O→O P→P relationships I was making a rather noticable
subset of the fbx dataset and that seemed sad.

This way we should get the whole thing :)